Convert to using arguments
When developing on windows environment variables have to be specifically set
unlike *nix environments. This breaks the universal nature
of package.json.

Using process.argv variables we can check for a specific flag and if
present we can setup variables in our code accordingly.
In this case we want to set the dev server url if  passing --start-dev
via package.json scripts.